requirements.txt
django-simple-captcha==0.5.12